+++ ReleaseNotes.md
@@ -23,6 +23,8 @@ SOFTWARE.
 ## v3.7
 
 * Partial MISRA C2012 compliance
+* I_FIND_BINARY_HEADER_USEFUL symbol renamed to USE_INO_BINARY_HEADER
+* arm_typedefs & arm_chip_ino: fix for USE_INO_BINARY_HEADER (re-define B0 & B1)
 * arm_errors: error codes added
 * arm_chip_stm32: STM32_UNIQUE_ID split to get 3 different parts of unique ID
 * arm_inlines_float: get_fp_dec fix for negative values